Fixed the abstraction and data is showing correctly.
