+++ src/components/layout/footer/footer.css
@@ -41,4 +41,3 @@
   text-align: right;
   flex: 1; /* Allow elements to share space evenly */
 }
-
